Update pom.xml
Instead of relying on the BuildTools-only spigot artifact, we now rely on the publicly available spigot-api artifact. This is possible, since the API does not make use of any CraftBukkit or NMS code or any other code that is exclusively provided by the spigot artifact. The scope is also set to provided, so the artifact isn't needlessly shaded into the final jar. This change will help developers who want to make use of this API by ensuring the artifact can be fetched online, without having the run BuildTools first. Compatibility is still ensured, since the minimum version is still at 1.9.